+++ test/integration/index.js
@@ -4,6 +4,7 @@ require('./spec/connect');
 require('./spec/localparticipant');
 require('./spec/localtracks');
 require('./spec/localtrackpublication');
+require('./spec/rest');
 require('./spec/room');
 require('./spec/util/insightspublisher');
 require('./spec/util/support');