+++ test/unit/index.js
@@ -52,7 +52,7 @@ require('./spec/signaling/v2/room');
 require('./spec/signaling/v2/peerconnection');
 require('./spec/signaling/v2/peerconnectionmanager');
 require('./spec/signaling/v2/localtrackpublication');
-require('./spec/signaling/v2/remotetrack');
+require('./spec/signaling/v2/remotetrackpublication');
 require('./spec/signaling/v2/transport');
 
 require('./spec/util');